requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

annotated-types==0.7.0
certifi==2024.12.14
click==8.1.8
einops==0.8.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.27.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pillow==10.2.0
pydantic==2.10.6
pydantic-core==2.27.2
pyyaml==6.0.2
regex==2024.11.6
safetensors==0.5.2
setuptools==78.1.1
sqlite-vec==0.1.6
sympy==1.13.1
timm==1.0.14
tokenizers==0.21.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
transformers==4.48.1
typing-extensions==4.12.2
customtkinter==5.2.1
pyinstaller==6.3.0
